--- a/torch/nn/parallel/distributed.py
+++ b/torch/nn/parallel/distributed.py
@@ -189,11 +189,13 @@ class DistributedDataParallel(Module):
             module.train(mode)
 
     def _sync_params(self):
-        params = [p.data for p in self.module.parameters()]
-        result = broadcast_coalesced(params, self.device_ids, self.broadcast_bucket_size)
-        for tensors, module in zip(result[1:], self._module_copies[1:]):
-            for tensor, param in zip(tensors, module.parameters()):
-                param.data.set_(tensor)
+        if len(self.device_ids) > 1:
+            # intra-node parameter sync
+            params = [p.data for p in self.module.parameters()]
+            result = broadcast_coalesced(params, self.device_ids, self.broadcast_bucket_size)
+            for tensors, module in zip(result[1:], self._module_copies[1:]):
+                for tensor, param in zip(tensors, module.parameters()):
+                    param.data.set_(tensor)
 
         buffers = list(self.module._all_buffers())
         if len(buffers) > 0:
@@ -203,11 +205,12 @@ class DistributedDataParallel(Module):
             for buf, synced in zip(buffers, _unflatten_dense_tensors(flat_buffers, buffers)):
                 buf.copy_(synced)
 
-            # intra-node buffer sync
-            result = broadcast_coalesced(buffers, self.device_ids, self.broadcast_bucket_size)
-            for tensors, module in zip(result[1:], self._module_copies[1:]):
-                for tensor, buf in zip(tensors, module._all_buffers()):
-                    buf.set_(tensor)
+            if len(self.device_ids) > 1:
+                # intra-node buffer sync
+                result = broadcast_coalesced(buffers, self.device_ids, self.broadcast_bucket_size)
+                for tensors, module in zip(result[1:], self._module_copies[1:]):
+                    for tensor, buf in zip(tensors, module._all_buffers()):
+                        buf.set_(tensor)
 
     def _register_grad_hooks(self):
         self._grad_accs = []  # need to keep them in scope